requirements.txt
requests>=1,<3
python-dateutil>=2.1,<3
six>=1.3,<2
texttable>=0.8.6,<0.9
